Time: 58 ms (43.07%), Space: 78.1 MB (29.22%) - LeetHub

+++ 1386-cinema-seat-allocation/1386-cinema-seat-allocation.cpp
@@ -7,7 +7,7 @@ public:
             mp[x[0]].insert(x[1]);
         }
 
-        long long ans = 2LL * n;
+        int ans = 2LL * n;
 
         for (auto &[row, seats] : mp) {
